% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_labels.R
\name{create_github_labels}
\alias{create_github_labels}
\title{Create GitHub labels for issues and PR from CSV file}
\usage{
create_github_labels(label_csv, owner = "fmichonneau",
  repo = "test-repo-labels", delete_previous = FALSE)
}
\arguments{
\item{label_csv}{the path to the CSV file with the labels. It should have
the following columns: type, label, color, description, long_description.}

\item{owner}{the owner of the GitHub repository}

\item{repo}{the name of the GitHub repository where to create the labels}

\item{delete_previous}{should already existing labels in the repository be deleted? (boolean)}
}
\value{
\code{TRUE} invisibly
}
\description{
Labels that already exists (with the same name as the one provided in the
CSV file) will be updated to match the specification of the spreadsheet.
}
\details{
If \code{delete_previous} is set to \code{TRUE}, any label that were present
in the repository (and that do not match one of the labels specified in the
CSV file) will be deleted.
}
