% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{create_issue}
\alias{create_issue}
\alias{create_milestone}
\title{Issues and milestones}
\usage{
create_issue(owner, repo, title, body, labels = c("help wanted"), ...)

create_milestone(owner, repo, title, ...)
}
\arguments{
\item{owner}{the owner of the GitHub repository}

\item{repo}{the name of the GitHub repository}

\item{title}{the title of the issue or milestone}

\item{body}{the body of the issue}

\item{labels}{GitHub label to use with the issue}

\item{...}{additional parameters being passed to gh::gh}
}
\value{
JSON
}
\description{
Create issues and milestones in repositories
}
