% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{generate_zenodo_json}
\alias{generate_zenodo_json}
\title{Generate a JSON file compatible with Zenodo to record lesson releases}
\usage{
generate_zenodo_json(
  repos,
  local_path,
  editors_github,
  since = NULL,
  ignore = character(0)
)
}
\arguments{
\item{repos}{a data frame that has three columns:
- name the type of repository. "main" for the repository to credit, 
     "source" for the originating repository if "main" is a translation,
     "template" for the lesson template repository, depending on translation
- owner the github owner name
- repo the repository name}

\item{local_path}{path to the local version of the repository}

\item{editors_github}{the github username of the editors as it appears in AMY}

\item{since}{a date string in ISO 8601 format as a start date to gather 
commits.}
}
\description{
Generate a JSON file compatible with Zenodo to record lesson releases
}
\note{
This is the main driver function for generating the ZENODO json file. 
  I think one thing that is not immediately clear is that "editors" in 
  zenodo terminology is the equivalent of our "maintainers".
}
