% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{gh_list_files_raw}
\alias{gh_list_files_raw}
\alias{gh_list_files}
\title{GitHub files}
\usage{
gh_list_files_raw(owner, repo, path = NULL, ...)

gh_list_files(owner, repo, path = NULL, ...)
}
\arguments{
\item{owner}{user or organization for the repository}

\item{repo}{name of the repository}

\item{path}{optional sub-folder within the repository}

\item{...}{additional parameters to be passed to \code{gh} (e.g. \code{.token})}
}
\value{
\itemize{
\item \code{gh_list_files_raw} returns a list (default from \code{gh}).
\item \code{gh_list_files} returns a tibble.
}
}
\description{
Gets the list of files from a GitHub repository.
}
\details{
By default, it only returns the files in the root of the repository. You can
specify a \code{path} to get the content of a sub-folder. There is no way to get
all content from a repository at this stage.

\code{gh_list_files} returns the name of the file, its path, and the URL for the
HTML rendering of the file.
}
