% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_labels.R
\name{list_all_github_labels}
\alias{list_all_github_labels}
\alias{delete_all_github_labels}
\title{List and delete GitHub labels}
\usage{
list_all_github_labels(owner, repo)

delete_all_github_labels(owner, repo)
}
\arguments{
\item{owner}{the owner of the GitHub repository}

\item{repo}{the name of the GitHub repository}
}
\description{
List and delete all GitHub labels from a repository.
}
