% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_labels.R
\name{summarize_github_labels}
\alias{summarize_github_labels}
\alias{document_github_labels}
\title{Summary of GitHub labels from CSV file}
\usage{
summarize_github_labels(label_csv, format = "html", escape = FALSE,
  ...)

document_github_labels(label_csv, out = NULL)
}
\arguments{
\item{label_csv}{the path to a CSV file that describes the labels. It should
have the following columns: print_order, type, label, color, use_prefix,
description, long_description. The canonical version of this file lives
in The Carpentries repository's
(\url{https://github.com/carpentries/handbook/}) \code{data/} folder.}

\item{format}{format of the output table (passed to \code{kable}).}

\item{escape}{should the HTML be escaped (passed to \code{kable}).}

\item{...}{additional arguments to be passed to \code{kable}.}

\item{out}{path to file to write the HTML file documenting the labels}
}
\value{
An HTML formatted table that shows the type of label, its color and
its description.
}
\description{
Creates an HTML formatted table or list to be included in the Carpentry
handbook or other resources
}
\details{
\code{summarize_github_labels} creates a table, while
\code{document_github_labels} creates a list.

See \code{\link[=create_github_labels]{create_github_labels()}} for a description of the CSV file format
expected by this function.
}
