% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_issue_per_file.R
\name{one_issue_per_file}
\alias{one_issue_per_file}
\title{One issue per file}
\usage{
one_issue_per_file(owner, repo, path = "_episodes", pattern = "md$",
  issue_title = "Translation of {name}", issue_body, ...)
}
\arguments{
\item{owner}{user or organization who owns the GitHub repository}

\item{repo}{name of the GitHub repository}

\item{path}{subfolder to look for files}

\item{pattern}{regular expression to select the files}

\item{issue_title}{string to use for the title of the issue}

\item{issue_body}{string to use for the content of the issue}

\item{...}{additional parameters to be passed to \code{gh}}
}
\description{
Opens one issue for each file found in a GitHub repository
}
