% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_labels.R
\name{summarize_github_labels}
\alias{summarize_github_labels}
\title{Summary table of GitHub labels from CSV file}
\usage{
summarize_github_labels(label_csv, format = "html", escape = FALSE, ...)
}
\arguments{
\item{label_csv}{the path to a CSV file that describes the labels. It should
have the following columns: type, label, color, description,
long_description.}

\item{format}{format of the output table (passed to \code{kable}).}

\item{escape}{should the HTML be escaped (passed to \code{kable}).}

\item{...}{additional arguments to be passed to `kable`.}
}
\value{
An HTML formatted table that shows the type of label, its color and
    its description.
}
\description{
Creates an HTML formatted table to be included in the Carpentry handbook or other resources
}
